pyproject.toml
[build-system]
requires = ["flit_core>=3.9.0"]
build-backend = "flit_core.buildapi"

[project]
name = "vilcos"
version = "0.1.2"
description = "A FastAPI-based web framework"
authors = [{name = "Nidal Alhariri", email = "level09@gmail.com"}]
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "jinja2>=3.1.2",
    "sqlalchemy>=2.0.23",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "starlette-session>=0.4.3",
    "aioredis>=2.0.1",
    "greenlet>=3.0.1",
    "asyncpg>=0.29.0",
    "typer>=0.9.0",
    "IPython>=8.17.2",
    "redis>=5.0.1",
    "argon2-cffi>=23.1.0",
]

[project.scripts]
vilcos = "vilcos.cli:main"

[project.urls]
Home = "https://github.com/level09/vilcos"
Source = "https://github.com/level09/vilcos"